requirements.txt
selenium
pytest
pytest-html
allure-pytest
openpyxl
pandas
webdriver_manager
